refactor(storage): reorg writer code
